refactor(session): extract WindowManagerMixin to _session_windows.py
Behavior-preserving move of register/unregister_plot, next_window_id,
_plot_by_window_id, and the close/forget/resize/figure-event cluster.
MDIManager fold DEFERRED (Session keeps self.mdi_manager).

Mixins call ipc.emit (not a module-bound emit) so the conftest emit
monkeypatch reaches them; same fix applied to the axes/testharness mixins.
